refactor(test): split network message builders
