+++ test/src/specs/relay/compact_block.rs
@@ -1,7 +1,8 @@
-use crate::utils::{
+use crate::util::message::{
     build_block, build_block_transactions, build_compact_block, build_compact_block_with_prefilled,
-    build_header, build_headers, clear_messages, wait_until,
+    build_header, build_headers,
 };
+use crate::utils::{clear_messages, wait_until};
 use crate::{Net, Spec, TestProtocol, DEFAULT_TX_PROPOSAL_WINDOW};
 use ckb_dao::DaoCalculator;
 use ckb_network::{bytes::Bytes, SupportProtocols};
